// packages/declare/src/generated/program.ts
export const ProgramBase_2_40 = z.object({
  accessLevel: ProgramAccessLevel_2_40.optional(),
  categoryCombo: refSchema('CategoryCombo'),
  code: z.string().max(50).optional(),
  completeEventsExpiryDays: z.number().int().optional(),
  description: z.string().optional(),
  displayFrontPageList: z.boolean().optional(),
  displayIncidentDate: z.boolean().optional(),
  enrollmentDateLabel: z.string().optional(),
  expiryDays: z.number().int().optional(),
  featureType: FeatureType_2_40.optional(),
  formName: z.string().optional(),
  ignoreOverdueEvents: z.boolean().optional(),
  incidentDateLabel: z.string().optional(),
  maxTeiCountToReturn: z.number().int().optional(),
  minAttributesRequiredToSearch: z.number().int().optional(),
  name: z.string().max(230),
  onlyEnrollOnce: z.boolean().optional(),
  openDaysAfterCoEndDate: z.number().int().optional(),
  organisationUnits: z.array(refSchema('OrganisationUnit')).optional(),
  programType: ProgramType_2_40,
  relatedProgram: refSchema('Program').optional(),
  selectEnrollmentDatesInFuture: z.boolean().optional(),
  selectIncidentDatesInFuture: z.boolean().optional(),
  shortName: z.string().max(50),
  skipOffline: z.boolean(),
  trackedEntityType: refSchema('TrackedEntityType').optional(),
  useFirstStageDuringRegistration: z.boolean().optional(),
  version: z.number().int().optional(),
})

export const ProgramBase_2_41 = z.object({
  accessLevel: ProgramAccessLevel_2_41.optional(),
  categoryCombo: refSchema('CategoryCombo'),
  code: z.string().max(50).optional(),
  completeEventsExpiryDays: z.number().int().optional(),
  description: z.string().optional(),
  displayFrontPageList: z.boolean().optional(),
  displayIncidentDate: z.boolean().optional(),
  enrollmentDateLabel: z.string().optional(),
  enrollmentLabel: z.string().optional(),
  eventLabel: z.string().optional(),
  expiryDays: z.number().int().optional(),
  featureType: FeatureType_2_41.optional(),
  followUpLabel: z.string().optional(),
  formName: z.string().optional(),
  ignoreOverdueEvents: z.boolean().optional(),
  incidentDateLabel: z.string().optional(),
  maxTeiCountToReturn: z.number().int().optional(),
  minAttributesRequiredToSearch: z.number().int().optional(),
  name: z.string().max(230),
  noteLabel: z.string().optional(),
  onlyEnrollOnce: z.boolean().optional(),
  openDaysAfterCoEndDate: z.number().int().optional(),
  organisationUnits: z.array(refSchema('OrganisationUnit')).optional(),
  orgUnitLabel: z.string().optional(),
  programStageLabel: z.string().optional(),
  programType: ProgramType_2_41,
  relatedProgram: refSchema('Program').optional(),
  relationshipLabel: z.string().optional(),
  selectEnrollmentDatesInFuture: z.boolean().optional(),
  selectIncidentDatesInFuture: z.boolean().optional(),
  shortName: z.string().max(50),
  skipOffline: z.boolean(),
  trackedEntityAttributeLabel: z.string().optional(),
  trackedEntityType: refSchema('TrackedEntityType').optional(),
  useFirstStageDuringRegistration: z.boolean().optional(),
  version: z.number().int().optional(),
})

export const ProgramBase_2_42 = z.object({
  accessLevel: ProgramAccessLevel_2_42.optional(),
  categoryCombo: refSchema('CategoryCombo'),
  code: z.string().max(50).optional(),
  completeEventsExpiryDays: z.number().int().optional(),
  description: z.string().max(255).optional(),
  displayFrontPageList: z.boolean().optional(),
  displayIncidentDate: z.boolean().optional(),
  enableChangeLog: z.boolean().optional(),
  enrollmentCategoryCombo: refSchema('CategoryCombo'),
  enrollmentDateLabel: z.string().max(255).optional(),
  enrollmentLabel: z.string().max(255).optional(),
  enrollmentsLabel: z.string().max(255).optional(),
  eventLabel: z.string().max(255).optional(),
  eventsLabel: z.string().max(255).optional(),
  expiryDays: z.number().int().optional(),
  featureType: FeatureType_2_42.optional(),
  followUpLabel: z.string().max(255).optional(),
  formName: z.string().max(255).optional(),
  ignoreOverdueEvents: z.boolean().optional(),
  incidentDateLabel: z.string().max(255).optional(),
  maxTeiCountToReturn: z.number().int().optional(),
  minAttributesRequiredToSearch: z.number().int().optional(),
  name: z.string().max(230),
  noteLabel: z.string().max(255).optional(),
  onlyEnrollOnce: z.boolean().optional(),
  openDaysAfterCoEndDate: z.number().int().optional(),
  organisationUnits: z.array(refSchema('OrganisationUnit')).optional(),
  orgUnitLabel: z.string().max(255).optional(),
  programStageLabel: z.string().max(255).optional(),
  programStagesLabel: z.string().max(255).optional(),
  programType: ProgramType_2_42,
  relatedProgram: refSchema('Program').optional(),
  relationshipLabel: z.string().max(255).optional(),
  selectEnrollmentDatesInFuture: z.boolean().optional(),
  selectIncidentDatesInFuture: z.boolean().optional(),
  shortName: z.string().max(50),
  skipOffline: z.boolean(),
  trackedEntityAttributeLabel: z.string().max(255).optional(),
  trackedEntityType: refSchema('TrackedEntityType').optional(),
  useFirstStageDuringRegistration: z.boolean().optional(),
  version: z.number().int().optional(),
})

export const ProgramBaseByTarget = {
  '2.40': ProgramBase_2_40,
  '2.41': ProgramBase_2_41,
  '2.42': ProgramBase_2_42,
} as const